allow starting multiple different processes at the same time
This will allow applications for which data might flow from one function
to another to process through the stream of data across multiple
processes with one command.  Rather than run 1 function to load up a
queue for the second function to read from, then running the second
function, and so on,  you can just run:

      `pytchfork <package>.[bin] <command1> <command2> <command3> `

To start up each process at once, and the flow of data from one to the
other will be immediate.  This is still a WIP, but this can be a quite
useful timesaving feature so I'd like to find the cleanest way to handle
this use case, along with the existing ones.